Emergency save of dirty Buffers when they are destroyed. Better fix to be committed for 1.6.5.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_1_6_X@31117 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/Buffer.cpp
@@ -3,7 +3,7 @@
  * This file is part of LyX, the document processor.
  * Licence details can be found in the file COPYING.
  *
- * \author Lars Gullik Bjønnes
+ * \author Lars Gullik Bjï¿½nnes
  * \author Stefan Schimanski
  *
  * Full author contact details are available in file CREDITS.
@@ -312,6 +312,20 @@ Buffer::~Buffer()
 	// clear references to children in macro tables
 	d->children_positions.clear();
 	d->position_to_children.clear();
+
+	if (!isClean()) {
+		docstring const text = bformat(_("The document %1$s has unsaved changes."
+				"\n\nDo you want to save the document or discard the changes?"), from_utf8(absFileName()));
+		int const ret = Alert::prompt(_("Save changed document?"),
+				text, 0, 2, _("&Save"), _("&Discard"));
+		switch (ret) {
+		case 0:
+			save();
+			break;
+		case 1:
+			break;
+		}
+	}
 
 	if (!d->temppath.destroyDirectory()) {
 		Alert::warning(_("Could not remove temporary directory"),
